Move markup decl and x-form to tree ns
Move the markup declaration processing, and its associated
transformation into is own namespace.  This includes changing
declaration transformation into a single stream of tags, primitives,
pseudo-tags and functions.  Each tag representation specifies the tag,
its attributes, and the number of children in the branch that follows.

The original tree test, which applied to functions in the core namespace
has been re-purposed to test the tree functions, as well as been made a
cljc file, so that it can be run on any platform.

Signed-off-by: Peter Schwarz <peter.schwarz@bicycle.io>