Implement render logic using diffed tree seq
Implement the render logic using the new tree sequence format, after a
diff has been applied.  This new renderer applies the changes in a
single loop.  Its sequential nature also allows for each node to be
visited at creation or destroy time.  The render code in the core
namespace did not allow this case. In order to support nested
on-mount/unmount functions, this style of rendering was required.

Note, as of this commit, pseudo-tags are not yet properly supported.

Signed-off-by: Peter Schwarz <peter.schwarz@bicycle.io>